refactor: 삭제, 상세 보기 기능 리액트 쿼리로 리팩터링
